py/docker/app-linux-amd64/requirements.txt
numpy>=1.21.5<2.0
scikit-rf>=0.18.1<1.0
websocket-client>=1.2.3<2.0


